slight changes in Matrix class

+++ Tools/Matrix.h
@@ -9,16 +9,16 @@ using namespace std;
 //template <class Type>
 class Matrix {
 private:
-    size_t _rows;
-    size_t _columns;
+    int _rows;
+    int _columns;
     //unique_ptr<scalar[]> data; // c++14 and higher
     vector<scalar> data;
 public:
     Matrix() : _rows(0), _columns(0) {};
     Matrix(size_t rows, size_t columns);
     Matrix(size_t rows, size_t columns, scalar value);
-    size_t rows() const;
-    size_t columns() const;
+    int rows() const;
+    int columns() const;
     scalar* data_ptr();
     scalar& operator()(size_t row, size_t column);
     scalar operator()(size_t row, size_t column) const;
